# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VflBlackJack API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Player {
  id: number;
  name: string;
  /** Current Jeton value (chip purchases only, excludes fixum) */
  chipBalance: number;
  /** Total fixum paid (5€ per session joined + registration) */
  fixumPaid: number;
  createdAt: string;
}

export interface PlayerInput {
  /** @minLength 1 */
  name: string;
}

export interface BuyChipsInput {
  /** @minimum 0.01 */
  amount: number;
}

export interface CashoutResult {
  player: Player;
  cashoutAmount: number;
  newBankBalance: number;
}

export interface BalanceSnapshot {
  id: number;
  sessionId: number;
  sessionName: string;
  balanceBefore: number;
  balanceAfter: number;
  difference: number;
  capturedAt: string;
}

export interface Bank {
  balance: number;
}

export interface Stats {
  bankBalance: number;
  playerCount: number;
  totalChipsInPlay: number;
}

export type GameSessionStatus =
  (typeof GameSessionStatus)[keyof typeof GameSessionStatus];

export const GameSessionStatus = {
  active: "active",
  ended: "ended",
} as const;

export interface GameSession {
  id: number;
  name: string;
  status: GameSessionStatus;
  createdAt: string;
  endedAt?: string | null;
  playerCount: number;
}

export interface GameSessionInput {
  /** @minLength 1 */
  name: string;
}

export type GameSessionDetailStatus =
  (typeof GameSessionDetailStatus)[keyof typeof GameSessionDetailStatus];

export const GameSessionDetailStatus = {
  active: "active",
  ended: "ended",
} as const;

export interface SessionPlayer {
  sessionPlayerId: number;
  playerId: number;
  name: string;
  chipBalance: number;
  fixumPaid: number;
  joinedAt: string;
}

export interface GameSessionDetail {
  id: number;
  name: string;
  status: GameSessionDetailStatus;
  createdAt: string;
  endedAt?: string | null;
  players: SessionPlayer[];
}

export interface AddPlayerToSessionInput {
  playerId: number;
}

export interface RemovePlayerResult {
  success: boolean;
  revertedAmount: number;
}

export interface PlayerFinalBalance {
  playerId: number;
  /** @minimum 0 */
  finalBalance: number;
}

export interface FinalizeSessionInput {
  balances: PlayerFinalBalance[];
}