# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VflBlackJack API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: players
  - name: bank
  - name: sessions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players
      responses:
        "200":
          description: List of all players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Create a new player (5€ Registrierungs-Fixum goes to bank)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "201":
          description: Created player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"

  /players/{id}:
    delete:
      operationId: deletePlayer
      tags: [players]
      summary: Delete and cashout a player (chip balance deducted from bank)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted player with cashout info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashoutResult"
        "404":
          description: Player not found

  /players/{id}/buy-chips:
    post:
      operationId: buyChips
      tags: [players]
      summary: Player buys additional chips
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyChipsInput"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Player not found

  /players/{id}/history:
    get:
      operationId: getPlayerHistory
      tags: [players]
      summary: Get balance history for a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of balance snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BalanceSnapshot"
        "404":
          description: Player not found

  /bank:
    get:
      operationId: getBank
      tags: [bank]
      summary: Get current bank balance
      responses:
        "200":
          description: Bank status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bank"

  /stats:
    get:
      operationId: getStats
      tags: [bank]
      summary: Get overview statistics
      responses:
        "200":
          description: App-wide stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

  /game-sessions:
    get:
      operationId: listGameSessions
      tags: [sessions]
      summary: List all game sessions
      responses:
        "200":
          description: List of game sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameSession"
    post:
      operationId: createGameSession
      tags: [sessions]
      summary: Create a new game session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameSessionInput"
      responses:
        "201":
          description: Created game session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSession"

  /game-sessions/active:
    get:
      operationId: getActiveSession
      tags: [sessions]
      summary: Get the currently active game session with its players
      responses:
        "200":
          description: Active session or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSessionDetail"

  /game-sessions/{id}:
    get:
      operationId: getGameSession
      tags: [sessions]
      summary: Get a game session with its players
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game session detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSessionDetail"
        "404":
          description: Not found

  /game-sessions/{id}/end:
    post:
      operationId: endGameSession
      tags: [sessions]
      summary: End a game session without finalizing balances
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ended session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSession"
        "404":
          description: Not found

  /game-sessions/{id}/finalize:
    post:
      operationId: finalizeGameSession
      tags: [sessions]
      summary: Finalize a game session with final chip balances for each player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinalizeSessionInput"
      responses:
        "200":
          description: Finalized session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSession"
        "404":
          description: Not found

  /game-sessions/{id}/players:
    post:
      operationId: addPlayerToSession
      tags: [sessions]
      summary: Add a player to a game session (5€ Fixum applied)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPlayerToSessionInput"
      responses:
        "201":
          description: Player added to session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionPlayer"
        "404":
          description: Not found
        "409":
          description: Player already in session

  /game-sessions/{id}/players/{playerId}:
    delete:
      operationId: removePlayerFromSession
      tags: [sessions]
      summary: Remove a player from a game session (reverts Fixum)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemovePlayerResult"
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Player:
      type: object
      required: [id, name, chipBalance, fixumPaid, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        chipBalance:
          type: number
          description: Current Jeton value (chip purchases only, excludes fixum)
        fixumPaid:
          type: number
          description: Total fixum paid (5€ per session joined + registration)
        createdAt:
          type: string

    PlayerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1

    BuyChipsInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          minimum: 0.01

    CashoutResult:
      type: object
      required: [player, cashoutAmount, newBankBalance]
      properties:
        player:
          $ref: "#/components/schemas/Player"
        cashoutAmount:
          type: number
        newBankBalance:
          type: number

    BalanceSnapshot:
      type: object
      required: [id, sessionId, sessionName, balanceBefore, balanceAfter, difference, capturedAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        sessionName:
          type: string
        balanceBefore:
          type: number
        balanceAfter:
          type: number
        difference:
          type: number
        capturedAt:
          type: string

    Bank:
      type: object
      required: [balance]
      properties:
        balance:
          type: number

    Stats:
      type: object
      required: [bankBalance, playerCount, totalChipsInPlay]
      properties:
        bankBalance:
          type: number
        playerCount:
          type: integer
        totalChipsInPlay:
          type: number

    GameSession:
      type: object
      required: [id, name, status, createdAt, playerCount]
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [active, ended]
        createdAt:
          type: string
        endedAt:
          type: string
          nullable: true
        playerCount:
          type: integer

    GameSessionInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1

    GameSessionDetail:
      type: object
      required: [id, name, status, createdAt, players]
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [active, ended]
        createdAt:
          type: string
        endedAt:
          type: string
          nullable: true
        players:
          type: array
          items:
            $ref: "#/components/schemas/SessionPlayer"

    SessionPlayer:
      type: object
      required: [sessionPlayerId, playerId, name, chipBalance, fixumPaid, joinedAt]
      properties:
        sessionPlayerId:
          type: integer
        playerId:
          type: integer
        name:
          type: string
        chipBalance:
          type: number
        fixumPaid:
          type: number
        joinedAt:
          type: string

    AddPlayerToSessionInput:
      type: object
      required: [playerId]
      properties:
        playerId:
          type: integer

    RemovePlayerResult:
      type: object
      required: [success, revertedAmount]
      properties:
        success:
          type: boolean
        revertedAmount:
          type: number

    FinalizeSessionInput:
      type: object
      required: [balances]
      properties:
        balances:
          type: array
          items:
            $ref: "#/components/schemas/PlayerFinalBalance"

    PlayerFinalBalance:
      type: object
      required: [playerId, finalBalance]
      properties:
        playerId:
          type: integer
        finalBalance:
          type: number
          minimum: 0