# lib/api-zod/src/generated/types/balanceSnapshot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VflBlackJack API
 * OpenAPI spec version: 0.1.0
 */

export interface BalanceSnapshot {
  id: number;
  sessionId: number;
  sessionName: string;
  balanceBefore: number;
  balanceAfter: number;
  difference: number;
  capturedAt: string;
}