# lib/api-zod/src/generated/types/cashoutResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VflBlackJack API
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from "./player";

export interface CashoutResult {
  player: Player;
  cashoutAmount: number;
  newBankBalance: number;
}