# lib/api-zod/src/generated/types/finalizeSessionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VflBlackJack API
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerFinalBalance } from "./playerFinalBalance";

export interface FinalizeSessionInput {
  balances: PlayerFinalBalance[];
}