# lib/api-zod/src/generated/types/gameSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VflBlackJack API
 * OpenAPI spec version: 0.1.0
 */
import type { GameSessionStatus } from "./gameSessionStatus";

export interface GameSession {
  id: number;
  name: string;
  status: GameSessionStatus;
  createdAt: string;
  endedAt?: string | null;
  playerCount: number;
}