# lib/api-zod/src/generated/types/gameSessionDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VflBlackJack API
 * OpenAPI spec version: 0.1.0
 */
import type { GameSessionDetailStatus } from "./gameSessionDetailStatus";
import type { SessionPlayer } from "./sessionPlayer";

export interface GameSessionDetail {
  id: number;
  name: string;
  status: GameSessionDetailStatus;
  createdAt: string;
  endedAt?: string | null;
  players: SessionPlayer[];
}