# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VflBlackJack API
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: number;
  name: string;
  /** Current Jeton value (chip purchases only, excludes fixum) */
  chipBalance: number;
  /** Total fixum paid (5€ per session joined + registration) */
  fixumPaid: number;
  createdAt: string;
}