# lib/api-zod/src/generated/types/playerFinalBalance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VflBlackJack API
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerFinalBalance {
  playerId: number;
  /** @minimum 0 */
  finalBalance: number;
}