# lib/api-zod/src/generated/types/sessionPlayer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VflBlackJack API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionPlayer {
  sessionPlayerId: number;
  playerId: number;
  name: string;
  chipBalance: number;
  fixumPaid: number;
  joinedAt: string;
}