# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VflBlackJack API
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  bankBalance: number;
  playerCount: number;
  totalChipsInPlay: number;
}